+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fixed an error in new science support for DMOS (Sir Mortimer)
 * Breaking Grounds DLC: Fixed the science value indication when the transmission is complete (Sir Mortimer)
 * Crediting of zero science is now reduced to a minimum. This will fix the constant science updates with mods like Strategia (Sir Mortimer)
+* Added mod support to query the radiation model, as well as toggle visibility of belts and magnetopause (Sir Mortimer)
 
 ## v3.0.2 for all versions of KSP from 1.4.0 to 1.7.x